Merged PR 26648: updated to GizmoSDK 2.12.230.1
updated to GizmoSDK 2.12.230.1

A breaking bug was introduced in GizmoSDK 2.12.189.1 where the NodeActionEvent enum diverged between native and managed gizmo. To avoid having to change gizmo a hotfix was added where we correct for the enum value diff by manually adding 2. This fix will be reverted once we update to a later gizmo version where this is fixed.

Related work items: #305922